Prepare for release on Bower
To prepare for a release on Bower, added bower.json describing
the project. In the future, if new layers are added, they can
be added to the main list.

This is only for the 1.0.0-rc.1 version of Leaflet, as specified
in the dependencies.

Also added a .gitignore to ignore the bower_components directory
(and also test directories while I'm at it).